requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap4==23.2
django-bootstrap5==23.3
django-crispy-forms==2.1
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
